Remove package_manifest attribute for packages
Use packages throughout with a new "filter" scan substage that calls
the process_codebase of a scan plugin after the scans. This used
by the package plugin to move manifest data to the root directory
of the package they document.

Add new Package "manifest_path" attribute which is a relative path to
the manifest file if any, such as a Maven .pom or a npm package.json.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>